Commonizing test functions
