+++ src/lint/rules/test/mod.rs
@@ -7,30 +7,9 @@ pub mod common;
 #[cfg(test)]
 pub mod indentation_tests {
 
-use crate::lint::LintCfg;
-use crate::lint::rules::tests::{run_linter, assert_snippet};
-use crate::lint::rules::instantiate_rules;
-use crate::lint::rules::CurrentRules;
-use crate::analysis::LocalDMLError;
+use crate::lint::rules::tests::{assert_snippet};
+use crate::lint::rules::test::common::{set_up, assert_indentation};
 
-fn set_up() -> CurrentRules {
-    let cfg = LintCfg::default();
-    instantiate_rules(&cfg)
-}
-
-fn assert_indentation(
-    code: &str, expected_errors: usize, rules: CurrentRules)
-{
-    let lint_errors = run_linter(code, &rules);
-    let Ok(ref lint_errors) = lint_errors else {
-        panic!();
-    };
-    let mut indent_errors: Vec<&LocalDMLError> = vec!();
-    for error in lint_errors {
-        indent_errors.push(error);
-    }
-    assert_eq!(indent_errors.len(), expected_errors, "{:#?}", lint_errors);
-}
 
 pub static IN3_FUNCTION_CONTENTS_CORRECT_INDENT: &str = "
 method some_function(int a) {
